Polar without test

+++ Libin_predict/OPT.cpp
@@ -44,7 +44,7 @@ vector<edge_t> E;
 vector<worker_t> workers;
 vector<task_t> tasks;
 int workerN, taskN;
-int dw, dr, vw;
+double dw, dr, vw;
 int slotN, gridLength, gridWidth;
 int st, ed;
 #ifdef WATCH_MEM
@@ -94,7 +94,7 @@ void init_network() {
 void init() {
 	vector<predictItem_t> predictItems;
 
-	scanf("%d %d %d %d %d %d %d %d", &workerN, &taskN, &dw, &dr, &vw, &slotN, &gridLength, &gridWidth);
+	scanf("%d %d %lf %lf %lf %d %d %d", &workerN, &taskN, &dw, &dr, &vw, &slotN, &gridLength, &gridWidth);
 	const int itemN = workerN + taskN;
 	int typeId;
 	worker_t worker;
